Updated Astaire to use a copy of the old ActionPack::Cascade
